Remove unneeded casts in Songbird.java Make G15WrapperWindows.java actually compile (It still does nothing, but it now compiles) Some work on G15WrapperLinuxNoComposer.java - Shape drawing is complete, just need text now and we can drop the G15Composer requirement :) G15Wrapper now uses sane values for LCD_WIDTH and LCD_HEIGHT Fix G15Control.java not allowing G15WrapperLinuxNoComposer to be used.

+++ trunk/src/uk/org/dataforce/g15/plugins/songbird/Songbird.java
@@ -193,8 +193,8 @@ public class Songbird implements Plugin {
 					int currentTime = 0;
 					int totalTime = 1;
 						try {
-						currentTime = (int)(Integer.parseInt(info.get(PLAYER_POSITION))/1000);
-						totalTime = (int)(Integer.parseInt(info.get(PLAYER_LENGTH))/1000);
+						currentTime = Integer.parseInt(info.get(PLAYER_POSITION))/1000;
+						totalTime = Integer.parseInt(info.get(PLAYER_LENGTH))/1000;
 						if (totalTime == 0) {
 							currentTime = 0;
 							totalTime = 1;
